backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.92.0",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "markdown2>=2.5.5",
    "mysqlclient>=2.2.8",
    "python-dotenv>=1.2.2",
]